Fixed directory issues for special characters

+++ mod-downloader.ps1
@@ -11,16 +11,20 @@
 function Get-ModsList {
     param(
         [Parameter()]
-        [string]$dest,
-        [Parameter()]
         [string]$source
     )
-
-    Remove-Item -Path ".\modslist.list"
-    $wc = New-Object System.Net.WebClient
-    $wc.DownloadFile($source, $dest)
-    Start-Sleep -Seconds 2
-    return test-path -Path $dest
+    
+    try{
+        if((Test-Path -path ".\modslist.list") -eq $true){
+            Remove-Item -Path ".\modslist.list"
+        }
+        Invoke-WebRequest -URI $source -OutFile ".\modslist.list"
+        Start-Sleep -Seconds 3
+    }
+    catch{
+        Write-Error $_.exception
+    }
+    return test-path -Path ".\modslist.list"
 }
 
 function Get-AModFromURI {
@@ -31,46 +35,47 @@ function Get-AModFromURI {
         [string]$target
     )
 
-    $name = ($source -split "/")
-    $name = $name[$name.Length -1] -replace "/",""
-    Write-Verbose "Working on mod: $name"
-    
-    if((test-path -path "$target\$name") -eq $false){
-        $wc = New-Object System.Net.WebClient
-        $wc.DownloadFile($source, "$target\$name")
-        Start-Sleep -Seconds 2
+    try{
+        $name = ($source -split "/")
+        $name = $name[$name.Length -1] -replace "/",""
+        Write-Verbose "Working on mod: $name"
+
+        if((test-path -path ".\mods\$name") -eq $false){
+            Invoke-WebRequest -URI $source -OutFile ".\mods\$name"
+            Start-Sleep -Seconds 4
+        }
     }
-    return test-path -Path "$target\$name"
+    catch{
+        Write-Error $_.exception
+    }
+    return (test-path -Path ".\mods\$name")
 }
 
-function Get-MyMods {
+function Invoke-ModDownloader {
     [CmdletBinding()]
     param()
 
     Clear-Host
     if((test-path -Path ".\config.cfg") -eq $false){
         $errorMsg = "Could not find config.cfg"
-
         new-item -Path ".\config.cfg"
-
         if((test-path -Path ".\config.cfg") -eq $false){
             $errorMsg += ", failed generating config file for you."
         }
         else{
-            $errorMsg += ", successfully created file for you. On the first row: enter where to save the mods, on the second row, enter URI to mod repository.`n`nExample:`nC:\temp\mods`nhttps://google.com"
+            $errorMsg += ", successfully created file for you. On the first row: enter URI to mod repository.`n`nExample:`nC:\temp\mods`nhttps://google.com"
         }
         Write-Error $errorMsg
         exit
     }
 
-    $modslist = ".\modslist.list"
     $config = Get-Content -Path ".\config.cfg"
 
     Write-Verbose "Downloading mod-list from $($config[1])"
 
-    if((Get-ModsList -dest $modslist -source $config[1]) -eq $true){
-        Get-Content -Path $modslist | ForEach-Object {
-            if((Get-AModFromURI -source $_ -target $config[0] -Verbose) -eq $false){
+    if((Get-ModsList -source $config) -eq $true){
+        Get-Content -Path ".\modslist.list" | ForEach-Object {
+            if((Get-AModFromURI -source $_) -eq $false){
                 Write-Error "Failed downloading $_"
             }
             else{
@@ -83,4 +88,4 @@ function Get-MyMods {
     }
 }
 
-Get-MyMods -Verbose
+Invoke-ModDownloader -Verbose
